Fix BTPaymentButton crash with tokenization key
BTPaymentButton would crash the demo app when tokenization key was
used. This was fixed by changing its initializer to call [super init]
instead of [super initWithFrame:CGRectZero]. The current hypothesis
is that the payment button would attempt to calculate its layout
attributes before it was rendered on screen, which causes a range
exception to be thrown in the custom flow layout.